added language & toast & improved

+++ api/config.ts
@@ -1,7 +1,7 @@
 import axios from "axios";
 import * as SecureStore from "expo-secure-store";
 
-const BASE_URL = process.env.EXPO_PUBLIC_API_URL;
+const BASE_URL = "http://172.20.10.3:3000/api";
 
 
 const apiClient = axios.create({
@@ -35,6 +35,13 @@ apiClient.interceptors.response.use(
     (response) => response,
     async (error) => {
         if (error.response?.status === 401) {
+
+            // 如果是登录/注册请求，直接把原始错误抛出，不做任何处理
+            const url = error.config?.url || '';
+            if (url.includes('/auth/login') || url.includes('/auth/signup') || url.includes('/auth/google')) {
+                return Promise.reject(error); // 保留原始 error，让页面自己处理
+            }
+
             try {
                 // 清除本地存储的过期 Token
                 await SecureStore.deleteItemAsync("user_token");
